requirements.txt
streamlit>=1.30.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
google-generativeai>=0.8.0
sqlalchemy>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pypdf>=4.0.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
